==> notebooks/02-Mudanca_Formato_Correcoes.ipynb <==
# Mudança para formato longo e algumas correcções básicas

## Inicialização

import numpy as np
import pandas as pd

# Diretoria onde os dados estão localizados
dir_original_files = './02-dados-qualar-integrados/'

# Directoria onde vamos guardar os resultados 
dir_destination = './03-dados-qualar-longo-corrigido/'

# ficheiro das medições 
ficheiro_medicoes = dir_original_files + 'df_estacoes.csv'

df = pd.read_csv(ficheiro_medicoes, thousands=',', index_col=0)

df

## Mudança de nomes de colunas
Vamos usar nomes mais simples:

    - Dióxido de Enxofre (µg/m3): SO2
    - Partículas < 10 µm (µg/m3): PM10
    - Ozono (µg/m3): O3
    - Dióxido de Azoto (µg/m3): NO2
    - Monóxido de Carbono (mg/m3): CO
    - Benzeno (µg/m3): Benzeno
    - Partículas < 2.5 µm (µg/m3): PM2.5

# Renomear as colunas atuais (esquerda), para os nomes mais simples (simples).
df.rename(columns={'Dióxido de Enxofre (µg/m3)' : 'SO2','Partículas < 10 µm (µg/m3)' : 'PM10',
                   'Dióxido de Azoto (µg/m3)' : 'NO2','Ozono (µg/m3)' : 'O3', 'Monóxido de Carbono (mg/m3)' : 'CO', 
                  'Benzeno (µg/m3)' : 'Benzeno', 'Partículas < 2.5 µm (µg/m3)' : 'PM2.5'},inplace = True)
# Conferir o resultado
df

# Como não iremos usar as medições de Benzendo e CO, podemos removê-las
df.drop(columns=['Benzeno','CO'],inplace=True)
# Conferir o resultado
df

## Converter para formato longo

# Cada linha será definida pela data (no index), pela estação e pelo poluente
df = df.melt(id_vars=['Estação'], ignore_index=False)
# Conferir o resultado
df

#Verificar existência de valores nulos nas medições
print('Existem valores nulos:')
print(df.isna().values.any())

# Agora podemos remover os nulos 
df.dropna(inplace = True)
# Conferir o resultado
print('Existem valores nulos:')
print(df.isna().values.any())
df

## Transformar os negativos em valores igual a 0
Devido à imprecisão dos sensores, por vezes são medidos valores negativos. Vamos mudar as ocorrências de valores negativos para 0

#Verificar a existência de valores negativos
negativos = df.loc[df['value']<0]
print('Existem valores negativos:')
print(not(negativos.empty))

# Mudar medições com valores negativos para 0
df['value'] = df['value'].apply(lambda x: 0 if x<0 else x)
# Conferir o resultado
negativos = df.loc[df['value']<0]
print('Existem valores negativos:')
print(not(negativos.empty))
df

## Guardar o novo CSV

# Guardar resultado em ficheiro csv
df.to_csv(dir_destination + '03-medicoes-longo.csv', index = True)

## Guardar o novo CSV com as medições de apenas as estações de AML (Área metropolitana de Lisboa - sul e norte)

Pode usar este subconjunto dos dados nas fases seguintes, caso o conjunto completo se prove grande demais

# Usando os dados das estações (estações.xlsx), filtrar as medições das estações de AML
# Dataframe regiões que contém a listagem das estações e regiões presentes nos dados
regioes = pd.read_excel(dir_original_files + 'estacoes.xlsx')
# Filtrar estações pelas quais estão presentes na região AML 
regioes = regioes[regioes['Região'].str.contains('AML')]
# Passar dataframe das estações de AML a lista
lista_regioes = regioes['Estação'].to_list()
# Filtrar dataframe pela lista de estação de AML
df_filtrado = df[df['Estação'].isin(lista_regioes)]
df_filtrado

# Guardar resultado em ficheiro csv
df_filtrado.to_csv(dir_destination + '03-medicoes-longo-AML.csv', index = True)
